Writing unit tests, fixing one bug, improving code
Wow, the simple fact of writing unit tests helped me found a bug and an
opportunity for improving the code.

What's left is writing a test for the function that reads 1 or 2 chars
depending on whether the first character is 'g'. To test that, I'd need
a stateful lambda/function in order to write a dummy generator of keys.

Unfortunately, if I've understood correcly, multi-statement lambdas
(that's how I'd write a stateful lambda) are not allowed before Vim9,
so I'm out of luck.